redid test page and changed file name to actually test
